V2.5.0 - Major Update
This major update brings much needed updates to many of the mods in the
modpack. It is also an update which will begin to move away from the
original Monster modpack this modpack is based on.

Also to note, several mods have been removed due to lack of usage on the
server or incompatibility, while several new mods have been added. You
will now be able to edit most parts of your world in singleplayer thanks
to WorldEdit.

Removed Mods:
- Agriculture
- BetterStorage
- CraftHeraldry
- MyTown (additionally removed aPerf & ForgePerms)
- OptiFine

Updated Mods:
- Damage Indicators
- CodeChickenLib
- ForgeMultipart
- AOBD
- BigReactors
- Blood Magic
- CoFHCore
- CodeChickenCore
- ComputerCraft
- DartCraft Beta
- DragonAPI
- DynamicTanks2
- EmasherResource
- EnderIO
- EngineersToolbox
- ExpandedRedstone
- ExtraCells
- Factorization
- Forge
- GeoStrata
- Hats
- JABBA
- LogisticsPipes
- LycanitesMobs
- Mariculture
- MeteorCraft (disabled by default)
- MineFactoryReloaded
- MobiusCore
- OpenBlocks
- OpenModsLib
- OpenPeripheralAddons
- OpenPeripheralCore
- Opis
- ProjectRedBase
- ProjectRedCompat
- ProjectRedIntegration
- ProjectRedLighting
- ProjectRedMechanical BETA
- ProjectRedWorld
- Random Things
- ReactorCraft
- TConstruct
- TMechworks
- Thaumcraft 4
- ThaumicTinkerer
- ThermalExpansion
- Backpacks
- bdlib
- compactsolars
- extrautils
- gendustry
- magical crops
- magicbees
- AdvancedSolarPanels
- GraviSuite
- mystcraft
- roguelike
- twilightforest

Newly Added Mods:
- CritterPet
- ElectriCraft
- Enchiridion
- EnderForest
- Forge NBTEdit
- WorldEdit-Forge
- LiteLoader